requirements.txt
anthropic==0.72.0
langchain-anthropic==1.0.2
langchain-core==1.0.4
python-dotenv==1.2.1
fastmcp==2.13.0.2
requests==2.32.3